+++ src/app/(PostPage)/[slug]/page.page.tsx
@@ -5,14 +5,14 @@ import {
   Note as PostNote,
   Share as PostShare,
 } from '@/app/(PostPage)/[slug]/components';
+import { PostSection } from '@/app/components/PostSection';
+import { TagSection } from '@/app/components/TagSection';
 import { Container } from '@/components/Functional/Container';
 import { Sidebar, Stack } from '@/components/UI/Layout';
 import { AUTHOR_NAME, READ_TIME_SUFFIX } from '@/constant';
 import { getBlogPostingStructured, getBreadcrumbStructured, getDescriptionText } from '@/lib/json-ld';
 import { getPostsJson } from '@/lib/posts';
 import { getOgpImage, getPermalink } from '@/lib/url';
-import { PostSection } from '@/pages/_components/PostSection';
-import { TagSection } from '@/pages/_components/TagSection';
 import type { Metadata } from 'next';
 import Script from 'next/script';
 import { getData } from './lib/getData';
