+++ src/app/components/TagSection/index.tsx
@@ -1,8 +1,8 @@
+import { LinkMore } from '@/app/components/home';
 import Heading from '@/components/UI/Heading';
 import { Box, Cluster } from '@/components/UI/Layout';
 import type { Props as PostTagProps } from '@/components/UI/Tag';
 import PostTag from '@/components/UI/Tag';
-import { LinkMore } from '@/pages/_components/home';
 import type { JSX } from 'react';
 
 type Props = {
