+++ src/app/libs/getDateAndUpdatedToSimpleFormat.test.ts
