+++ src/app/libs/getDateAndUpdatedToSimpleFormat.ts
